+++ front-api/lib/api/sse/sandbox_function_invocation_events.ts
@@ -1,4 +1,4 @@
-import { getSandboxFunctionInvocationEvents } from "@app/lib/api/sandbox/functions/events";
+import { getSandboxFunctionInvocationEvents } from "@app/lib/api/sandbox_functions/events";
 import type { Authenticator } from "@app/lib/auth";
 import { SandboxFunctionInvocationResource } from "@app/lib/resources/sandbox_function_invocation_resource";
 import { SandboxFunctionResource } from "@app/lib/resources/sandbox_function_resource";
