+++ front-api/routes/sse/w/[wId]/sandbox-functions/[functionId]/invocations/[invocationId]/events.test.ts
@@ -14,10 +14,10 @@ import {
 import type { JSONSchema7 as JSONSchema } from "json-schema";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
-vi.mock("@app/lib/api/sandbox/functions/events", async (importOriginal) => {
+vi.mock("@app/lib/api/sandbox_functions/events", async (importOriginal) => {
   const mod =
     await importOriginal<
-      typeof import("@app/lib/api/sandbox/functions/events")
+      typeof import("@app/lib/api/sandbox_functions/events")
     >();
   return {
     ...mod,
@@ -25,7 +25,7 @@ vi.mock("@app/lib/api/sandbox/functions/events", async (importOriginal) => {
   };
 });
 
-import { getSandboxFunctionInvocationEvents } from "@app/lib/api/sandbox/functions/events";
+import { getSandboxFunctionInvocationEvents } from "@app/lib/api/sandbox_functions/events";
 
 const inputSchema: JSONSchema = {
   type: "object",
