+++ front-api/routes/v1/w/[wId]/sandbox/sandbox-functions/result.test.ts
@@ -5,10 +5,10 @@ import {
 import { honoApp } from "@front-api/app";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
-vi.mock("@app/lib/api/sandbox/functions/events", async (importOriginal) => {
+vi.mock("@app/lib/api/sandbox_functions/events", async (importOriginal) => {
   const mod =
     await importOriginal<
-      typeof import("@app/lib/api/sandbox/functions/events")
+      typeof import("@app/lib/api/sandbox_functions/events")
     >();
   return {
     ...mod,
@@ -16,7 +16,7 @@ vi.mock("@app/lib/api/sandbox/functions/events", async (importOriginal) => {
   };
 });
 
-import { publishSandboxFunctionInvocationEvent } from "@app/lib/api/sandbox/functions/events";
+import { publishSandboxFunctionInvocationEvent } from "@app/lib/api/sandbox_functions/events";
 
 function postSandboxFunctionResult(
   workspace: { sId: string },
