+++ front-api/routes/v1/w/[wId]/sandbox/sandbox-functions/result.ts
@@ -1,5 +1,5 @@
 import { isSandboxFunctionInvocationTokenPayload } from "@app/lib/api/sandbox/access_tokens";
-import { publishSandboxFunctionInvocationEvent } from "@app/lib/api/sandbox/functions/events";
+import { publishSandboxFunctionInvocationEvent } from "@app/lib/api/sandbox_functions/events";
 import { sandboxApp } from "@front-api/middlewares/ctx";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 import { validate } from "@front-api/middlewares/validator";
